+++ tamr_toolbox/dataset/__init__.py
@@ -1,5 +1,5 @@
 """Tests for creating or modifying datasets in Tamr"""
 from . import manage
-from tamr_toolbox.dataset._profile import get_profile
+from tamr_toolbox.dataset._dataset import get_profile
 
 __all__ = ["manage", "get_profile"]
